Reorganise modules
- Remove import alias of concrete implementations of kodion classes
- Remove unused symbols from kodion module __all__ statement
- Merge kodion.ui.xbmc.info_labels into kodion.ui.xbmc.xbmc_items
  - Move combined kodion.ui.xbmc.xbmc_items to kodion.items.xbmc.xbmc_items
- Rename kodion.runner to kodion.plugin_runner
- Rename kodion.service to kodion.service_runner
- Move youtube_plugin.script_actions to kodion.script_actions
- Rename plugin.abstract_provider_runner.AbstractProviderRunner to plugin.abstract_plugin.AbstractPlugin
- Rename plugin.xbmc.xbmc_plugin.XbmcRunner to plugin.xbmc.xbmc_plugin.XbmcPlugin
