+++ resources/lib/plugin.py
@@ -11,8 +11,7 @@
 from __future__ import absolute_import, division, unicode_literals
 
 from youtube_plugin import youtube
-from youtube_plugin.kodion import runner
+from youtube_plugin.kodion import plugin_runner
 
 
-__provider__ = youtube.Provider()
-runner.run(__provider__)
+plugin_runner.run(youtube.Provider())
