+++ resources/lib/youtube_plugin/kodion/player/xbmc/xbmc_playlist.py
@@ -14,8 +14,7 @@ import json
 
 from ..abstract_playlist import AbstractPlaylist
 from ...compatibility import xbmc
-from ...items import VideoItem
-from ...ui.xbmc import xbmc_items
+from ...items import VideoItem, video_listitem
 
 
 class XbmcPlaylist(AbstractPlaylist):
@@ -33,7 +32,7 @@ class XbmcPlaylist(AbstractPlaylist):
         self._playlist.clear()
 
     def add(self, base_item):
-        uri, item, _ = xbmc_items.video_listitem(self._context, base_item)
+        uri, item, _ = video_listitem(self._context, base_item)
         if item:
             self._playlist.add(uri, listitem=item)
