+++ resources/lib/youtube_plugin/kodion/plugin/abstract_plugin.py
@@ -8,8 +8,10 @@
     See LICENSES/GPL-2.0-only for more information.
 """
 
+from __future__ import absolute_import, division, unicode_literals
 
-class AbstractProviderRunner(object):
+
+class AbstractPlugin(object):
     def __init__(self):
         pass
