+++ resources/lib/youtube_requests.py
@@ -12,7 +12,7 @@ from __future__ import absolute_import, division, unicode_literals
 import re
 
 from youtube_plugin.youtube.provider import Provider
-from youtube_plugin.kodion.context import Context
+from youtube_plugin.kodion.context import XbmcContext
 
 
 def __get_core_components(addon_id=None):
@@ -22,9 +22,9 @@ def __get_core_components(addon_id=None):
     """
     provider = Provider()
     if addon_id is not None:
-        context = Context(params={'addon_id': addon_id})
+        context = XbmcContext(params={'addon_id': addon_id})
     else:
-        context = Context()
+        context = XbmcContext()
     client = provider.get_client(context=context)
 
     return provider, context, client
