+++ resources/lib/youtube_resolver.py
@@ -12,15 +12,15 @@ from __future__ import absolute_import, division, unicode_literals
 import re
 
 from youtube_plugin.youtube.provider import Provider
-from youtube_plugin.kodion.context import Context
+from youtube_plugin.kodion.context import XbmcContext
 
 
 def _get_core_components(addon_id=None):
     provider = Provider()
     if addon_id is not None:
-        context = Context(params={'addon_id': addon_id})
+        context = XbmcContext(params={'addon_id': addon_id})
     else:
-        context = Context()
+        context = XbmcContext()
     client = provider.get_client(context=context)
 
     return provider, context, client
